move background notification modal logic outside
