+++ src/containers/App/App.tsx
@@ -1,40 +1,29 @@
-import React, { useCallback, useEffect } from "react";
+import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useHistory } from "react-router";
-import classNames from "classnames";
-import { Modal, TutorialModal } from "@/shared/components";
-import { BackgroundNotification } from "@/shared/components/BackgroundNotification";
-import { NotificationProvider } from "@/shared/components/Notification";
-import { useModal } from "@/shared/hooks";
-import { FirebaseCredentials } from "@/shared/interfaces/FirebaseCredentials";
-import { EventTypeState } from "@/shared/models/Notification";
-import { changeScreenSize, showNotification } from "@/shared/store/actions";
 import {
-  getNotification,
-  selectTutorialModalState,
-} from "@/shared/store/selectors";
-import { parseJson } from "@/shared/utils/json";
+  BackgroundNotificationModal,
+  TutorialModal,
+} from "@/shared/components";
+import { NotificationProvider } from "@/shared/components/Notification";
 import {
   ROUTE_PATHS,
   SMALL_SCREEN_BREAKPOINT,
   ScreenSize,
   WebviewActions,
-} from "../../shared/constants";
+} from "@/shared/constants";
+import { FirebaseCredentials } from "@/shared/interfaces/FirebaseCredentials";
+import { changeScreenSize } from "@/shared/store/actions";
+import { selectTutorialModalState } from "@/shared/store/selectors";
+import { parseJson } from "@/shared/utils/json";
 import { webviewLogin } from "../Auth/store/actions";
 import { TextDirectionHandler } from "./handlers";
 import { Router } from "./router";
 
 const App = () => {
   const dispatch = useDispatch();
-  const notification = useSelector(getNotification());
   const tutorialModalState = useSelector(selectTutorialModalState());
   const history = useHistory();
-
-  const {
-    isShowing: isShowingNotification,
-    onOpen: showNote,
-    onClose: closeNotification,
-  } = useModal(false);
 
   useEffect(() => {
     window.addEventListener("message", (event) => {
@@ -64,12 +53,6 @@ const App = () => {
   }, []);
 
   useEffect(() => {
-    if (notification) {
-      showNote();
-    }
-  }, [notification, showNote]);
-
-  useEffect(() => {
     const screenSize = window.matchMedia(
       `(min-width: ${SMALL_SCREEN_BREAKPOINT})`,
     );
@@ -97,38 +80,9 @@ const App = () => {
     };
   }, [dispatch]);
 
-  const closeNotificationHandler = useCallback(() => {
-    closeNotification();
-    dispatch(showNotification(null));
-    if (notification?.type === EventTypeState.fundingRequestAccepted) {
-      const path =
-        notification.additionalInformation === "0"
-          ? ROUTE_PATHS.PROPOSAL_DETAIL.replace(
-              ":id",
-              notification.eventObjectId,
-            )
-          : ROUTE_PATHS.SUBMIT_INVOICES.replace(
-              ":proposalId",
-              notification.eventObjectId,
-            );
-      history.push(path);
-    }
-  }, [closeNotification, history, notification, dispatch]);
-
   return (
     <>
-      {isShowingNotification && notification && (
-        <Modal
-          isShowing={isShowingNotification}
-          onClose={closeNotificationHandler}
-          className={classNames("notification")}
-        >
-          <BackgroundNotification
-            notification={notification}
-            closeHandler={closeNotificationHandler}
-          />
-        </Modal>
-      )}
+      <BackgroundNotificationModal />
       <TutorialModal isShowing={tutorialModalState.isShowing} />
       <NotificationProvider>
         <TextDirectionHandler />
