+++ src/shared/components/BackgroundNotificationModal/index.ts
@@ -1,0 +1,1 @@
+export { default as BackgroundNotificationModal } from "./BackgroundNotificationModal";
