+++ src/shared/components/index.tsx
@@ -40,3 +40,4 @@ export * from "./ProgressBar";
 export * from "./TutorialModal";
 export * from "./SupportShare";
 export * from "./Portal";
+export * from "./BackgroundNotificationModal";
